refactor: share Tree-sitter highlighting runtime
