+++ benchmarks/fixtures/example.go
