+++ benchmarks/fixtures/example.js
