+++ benchmarks/fixtures/example.rb
