+++ benchmarks/fixtures/example.rs
