+++ crates/lumis-cli/Cargo.toml
@@ -19,7 +19,7 @@ doc = false
 
 [dependencies]
 lumis-core = { path = "../lumis-core", version = "2", features = ["all-languages"] }
-lumis-wasm-runtime = { path = "../lumis-wasm-runtime" }
+lumis-wasm-runtime = { path = "../lumis-wasm-runtime", version = "0.1.0" }
 tree-sitter = { version = "^0.26.9", features = ["wasm"] }
 wasmtime = { version = "36", features = ["cache"] }
 streaming-iterator = "0.1"
