+++ crates/lumis-wasm-runtime/CHANGELOG.md
@@ -1,0 +1,1 @@
+<!-- Releases are prepended by `mise run release-prepare`. -->
