+++ packages/javascript/lumis/native/Cargo.toml
@@ -14,7 +14,7 @@ publish = false
 crate-type = ["cdylib"]
 
 [dependencies]
-lumis-wasm-runtime = { path = "../../../../crates/lumis-wasm-runtime" }
+lumis-wasm-runtime = { path = "../../../../crates/lumis-wasm-runtime", version = "0.1.0" }
 lumis-core = { path = "../../../../crates/lumis-core", version = "2", features = ["all-languages"] }
 napi = { version = "3", default-features = false, features = ["napi8"] }
 napi-derive = "3"
